Animations: Tunnel, Swipe, and Rainbow

+++ include/lights/led-buffer.h
@@ -17,13 +17,13 @@ namespace Lights
     LedBuffer(const LedBuffer &other) = delete;
     LedBuffer &operator=(const LedBuffer &other) = delete;
 
-    Color &operator()(uint8_t row, uint8_t column) { return leds[(row * Platform::Configuration::numColumns) + column]; }
-    const Color &operator()(uint8_t row, uint8_t column) const { return leds[(row * Platform::Configuration::numColumns) + column]; }
+    Color &operator()(int8_t row, int8_t column) { return leds[(row * Platform::Configuration::numColumns) + column]; }
+    const Color &operator()(int8_t row, int8_t column) const { return leds[(row * Platform::Configuration::numColumns) + column]; }
 
-    uint16_t size() const { return Platform::Configuration::numRows * Platform::Configuration::numColumns; }
+    int8_t size() const { return Platform::Configuration::numRows * Platform::Configuration::numColumns; }
 
-    Color &operator[](uint16_t index) { return leds[index]; }
-    const Color &operator[](uint16_t index) const { return leds[index]; }
+    Color &operator[](int8_t index) { return leds[index]; }
+    const Color &operator[](int8_t index) const { return leds[index]; }
 
     Color *data() { return leds; }
     const Color *data() const { return leds; }
